+++ Packages/UGF.Module.Assets/Runtime/AssetGroupAssetBase.cs
@@ -1,8 +1,0 @@
-﻿using UGF.Builder.Runtime;
-
-namespace UGF.Module.Assets.Runtime
-{
-    public abstract class AssetGroupAssetBase : BuilderAsset<IAssetGroup>
-    {
-    }
-}