+++ Packages/UGF.Module.Assets/Runtime/AssetLoadParameters.cs
@@ -1,21 +1,7 @@
-﻿using System;
-using UnityEngine;
-
-namespace UGF.Module.Assets.Runtime
+﻿namespace UGF.Module.Assets.Runtime
 {
-    [Serializable]
-    public struct AssetLoadParameters
+    public class AssetLoadParameters : IAssetLoadParameters
     {
-        [SerializeField] private AssetLoadMode m_mode;
-
-        public AssetLoadMode Mode { get { return m_mode; } set { m_mode = value; } }
-
-        public static AssetLoadParameters Default { get; } = new AssetLoadParameters(AssetLoadMode.Track);
-        public static AssetLoadParameters Direct { get; } = new AssetLoadParameters(AssetLoadMode.Direct);
-
-        public AssetLoadParameters(AssetLoadMode mode)
-        {
-            m_mode = mode;
-        }
+        public static AssetLoadParameters Empty { get; } = new AssetLoadParameters();
     }
 }