+++ Packages/UGF.Module.Assets/Runtime/AssetLoader`1.cs
@@ -1,0 +1,6 @@
+﻿namespace UGF.Module.Assets.Runtime
+{
+    public abstract class AssetLoader<TInfo> : AssetLoader<TInfo, IAssetLoadParameters, IAssetUnloadParameters> where TInfo : class, IAssetInfo
+    {
+    }
+}