+++ Packages/UGF.Module.Assets/Runtime/AssetTracker.cs
@@ -1,27 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
+using UGF.RuntimeTools.Runtime.Providers;
 
 namespace UGF.Module.Assets.Runtime
 {
-    public class AssetTracker : IAssetTracker
+    public class AssetTracker : Provider<string, AssetTrack>, IAssetTracker
     {
-        public IReadOnlyDictionary<string, AssetTrack> Tracks { get; }
-
-        private readonly Dictionary<string, AssetTrack> m_tracks = new Dictionary<string, AssetTrack>();
-
-        public AssetTracker()
-        {
-            Tracks = new ReadOnlyDictionary<string, AssetTrack>(m_tracks);
-        }
-
-        public bool Contains(string id)
-        {
-            if (string.IsNullOrEmpty(id)) throw new ArgumentException("Value cannot be null or empty.", nameof(id));
-
-            return m_tracks.ContainsKey(id);
-        }
-
         public AssetTrack Add(string id, object asset)
         {
             if (string.IsNullOrEmpty(id)) throw new ArgumentException("Value cannot be null or empty.", nameof(id));
@@ -29,21 +13,9 @@ namespace UGF.Module.Assets.Runtime
 
             var track = new AssetTrack(asset);
 
-            m_tracks.Add(id, track);
+            base.Add(id, track);
 
             return track;
-        }
-
-        public bool Remove(string id)
-        {
-            if (string.IsNullOrEmpty(id)) throw new ArgumentException("Value cannot be null or empty.", nameof(id));
-
-            return m_tracks.Remove(id);
-        }
-
-        public void Clear()
-        {
-            m_tracks.Clear();
         }
 
         public void Update(string id, AssetTrack track)
@@ -51,7 +23,8 @@ namespace UGF.Module.Assets.Runtime
             if (string.IsNullOrEmpty(id)) throw new ArgumentException("Value cannot be null or empty.", nameof(id));
             if (track.Asset == null) throw new ArgumentException($"Asset not found in specified track: '{track}'.");
 
-            m_tracks[id] = track;
+            Remove(id);
+            base.Add(id, track);
         }
 
         /// <summary>
@@ -86,14 +59,17 @@ namespace UGF.Module.Assets.Runtime
         /// UnTracks specified asset, decrements count when asset already tracked and removes when asset track count reached zero.
         /// </summary>
         /// <param name="id">The id of the asset.</param>
+        /// <param name="asset">The asset to untrack.</param>
         /// <param name="track">The modified asset track as result.</param>
         /// <returns>Returns True when asset track count reached zero and was removed, otherwise False.</returns>
-        public bool UnTrack(string id, out AssetTrack track)
+        public bool UnTrack(string id, object asset, out AssetTrack track)
         {
             if (string.IsNullOrEmpty(id)) throw new ArgumentException("Value cannot be null or empty.", nameof(id));
 
             if (TryGet(id, out track))
             {
+                if (track.Asset != asset) throw new ArgumentException($"Asset track not the same as specified asset: track:'{track.Asset}', asset:'{asset}'.");
+
                 track = Decrement(id);
 
                 if (track.Zero)
@@ -141,7 +117,7 @@ namespace UGF.Module.Assets.Runtime
         {
             if (asset == null) throw new ArgumentNullException(nameof(asset));
 
-            foreach (KeyValuePair<string, AssetTrack> pair in m_tracks)
+            foreach (KeyValuePair<string, AssetTrack> pair in this)
             {
                 if (pair.Value.Asset == asset)
                 {
@@ -152,23 +128,6 @@ namespace UGF.Module.Assets.Runtime
 
             track = default;
             return false;
-        }
-
-        public AssetTrack Get(string id)
-        {
-            return TryGet(id, out AssetTrack asset) ? asset : throw new ArgumentException($"Asset track not found by the specified id: '{id}'.");
-        }
-
-        public bool TryGet(string id, out AssetTrack track)
-        {
-            if (string.IsNullOrEmpty(id)) throw new ArgumentException("Value cannot be null or empty.", nameof(id));
-
-            return m_tracks.TryGetValue(id, out track);
-        }
-
-        public Dictionary<string, AssetTrack>.Enumerator GetEnumerator()
-        {
-            return m_tracks.GetEnumerator();
         }
     }
 }