+++ Packages/UGF.Module.Assets/Runtime/IAssetLoadParameters.cs
@@ -1,0 +1,6 @@
+﻿namespace UGF.Module.Assets.Runtime
+{
+    public interface IAssetLoadParameters
+    {
+    }
+}