+++ Packages/UGF.Module.Assets/Runtime/IAssetLoader.cs
@@ -1,13 +1,14 @@
 ﻿using System;
 using System.Threading.Tasks;
+using UGF.RuntimeTools.Runtime.Contexts;
 
 namespace UGF.Module.Assets.Runtime
 {
     public interface IAssetLoader
     {
-        object Load(IAssetProvider provider, string id, Type type);
-        Task<object> LoadAsync(IAssetProvider provider, string id, Type type);
-        void Unload(IAssetProvider provider, string id, object asset);
-        Task UnloadAsync(IAssetProvider provider, string id, object asset);
+        object Load(string id, Type type, IAssetLoadParameters parameters, IContext context);
+        Task<object> LoadAsync(string id, Type type, IAssetLoadParameters parameters, IContext context);
+        void Unload(string id, object asset, IAssetUnloadParameters parameters, IContext context);
+        Task UnloadAsync(string id, object asset, IAssetUnloadParameters parameters, IContext context);
     }
 }