+++ Packages/UGF.Module.Assets/Runtime/Loaders.Resources/ResourcesLoader.cs
@@ -1,27 +1,28 @@
 ﻿using System;
 using System.Threading.Tasks;
+using UGF.RuntimeTools.Runtime.Contexts;
 using UnityEngine;
 using Object = UnityEngine.Object;
 
 namespace UGF.Module.Assets.Runtime.Loaders.Resources
 {
-    public class ResourcesLoader : AssetLoader<IAssetGroup, IAssetInfo>
+    public class ResourcesLoader : AssetLoader<IAssetInfo>
     {
-        public bool ProviderAssetUnload { get; }
+        public bool EnableUnload { get; }
 
-        public ResourcesLoader(bool providerAssetUnload = true)
+        public ResourcesLoader(bool enableUnload = true)
         {
-            ProviderAssetUnload = providerAssetUnload;
+            EnableUnload = enableUnload;
         }
 
-        protected override object OnLoad(IAssetProvider provider, IAssetGroup group, IAssetInfo info, string id, Type type)
+        protected override object OnLoad(IAssetInfo info, string id, Type type, IAssetLoadParameters parameters, IContext context)
         {
             Object asset = UnityEngine.Resources.Load(info.Address, type);
 
             return asset ? asset : throw new NullReferenceException($"Resource load result is null by the specified arguments: id:'{id}', type:'{type}'.");
         }
 
-        protected override async Task<object> OnLoadAsync(IAssetProvider provider, IAssetGroup group, IAssetInfo info, string id, Type type)
+        protected override async Task<object> OnLoadAsync(IAssetInfo info, string id, Type type, IAssetLoadParameters parameters, IContext context)
         {
             ResourceRequest request = UnityEngine.Resources.LoadAsync(info.Address, type);
 
@@ -35,26 +36,29 @@ namespace UGF.Module.Assets.Runtime.Loaders.Resources
             return asset ? asset : throw new NullReferenceException($"Resource load result is null by the specified arguments: id:'{id}', type:'{type}'.");
         }
 
-        protected override void OnUnload(IAssetProvider provider, IAssetGroup group, IAssetInfo info, string id, object asset)
+        protected override void OnUnload(IAssetInfo info, string id, object asset, IAssetUnloadParameters parameters, IContext context)
         {
-            InternalUnload(asset);
+            if (EnableUnload)
+            {
+                InternalUnload(asset);
+            }
         }
 
-        protected override Task OnUnloadAsync(IAssetProvider provider, IAssetGroup group, IAssetInfo info, string id, object asset)
+        protected override Task OnUnloadAsync(IAssetInfo info, string id, object asset, IAssetUnloadParameters parameters, IContext context)
         {
-            InternalUnload(asset);
+            if (EnableUnload)
+            {
+                InternalUnload(asset);
+            }
 
             return Task.CompletedTask;
         }
 
-        private void InternalUnload(object asset)
+        private static void InternalUnload(object asset)
         {
-            if (ProviderAssetUnload)
-            {
-                if (!(asset is Object unityAsset)) throw new ArgumentException($"Asset must be a Unity Object to unload: '{asset}'.");
+            if (!(asset is Object unityAsset)) throw new ArgumentException($"Asset must be a Unity Object to unload: '{asset}'.");
 
-                UnityEngine.Resources.UnloadAsset(unityAsset);
-            }
+            UnityEngine.Resources.UnloadAsset(unityAsset);
         }
     }
 }