+++ Packages/UGF.Module.Assets/Runtime/Loaders.Resources/ResourcesLoaderAsset.cs
@@ -3,15 +3,15 @@
 namespace UGF.Module.Assets.Runtime.Loaders.Resources
 {
     [CreateAssetMenu(menuName = "Unity Game Framework/Assets/Resources Loader", order = 2000)]
-    public class ResourcesLoaderAsset : AssetLoaderAssetBase
+    public class ResourcesLoaderAsset : AssetLoaderAsset
     {
-        [SerializeField] private bool m_provideAssetUnload = true;
+        [SerializeField] private bool m_enableUnload = true;
 
-        public bool ProvideAssetUnload { get { return m_provideAssetUnload; } set { m_provideAssetUnload = value; } }
+        public bool EnableUnload { get { return m_enableUnload; } set { m_enableUnload = value; } }
 
         protected override IAssetLoader OnBuild()
         {
-            return new ResourcesLoader(m_provideAssetUnload);
+            return new ResourcesLoader(m_enableUnload);
         }
     }
 }